pyvr/moderngl_renderer/requirements.txt
moderngl
numpy
Pillow
scipy
matplotlib